backend/foodgram/requirements.txt
Django==4.1.5
djangorestframework
PyJWT==2.1.0
djoser==2.1.0

# djangorestframework-simplejwt==4.8.0